move director schunk driver to utils
