+++ apps/iiwa/iiwaManipApp.py
@@ -35,7 +35,7 @@ from PythonQt import QtCore, QtGui
 import drake as lcmdrake
 import bot_core as lcmbotcore
 
-from spartan.manipulation.director_schunk_driver import DirectorSchunkDriver
+from spartan.utils.director_schunk_driver import DirectorSchunkDriver
 import spartan.utils.director_utils as director_utils
 
 useROS = True
